+++ src/command_modules/azure-cli-resource/HISTORY.rst
@@ -2,6 +2,12 @@
 
 Release History
 ===============
+
+2.0.17
+++++++
+* `group export`: Fixed incompatibility with most recent version of msrest dependency.
+* `az policy assignment create`: policy assignment create command to work with built in policy definitions and policy set definitions.
+
 2.0.16 (2017-10-09)
 +++++++++++++++++++
 * group: permit --resource-group/-g options for resource group name.